refactor(manifest): split the parser into per-concern modules
The 1,533 lines left after the tests moved out were one flat file holding
twenty-two functions across three unrelated concerns. Now:

| Module | Lines | Parses |
|---|---|---|
| `manifest.rs` | 209 | `parse_manifest` itself, and the module wiring |
| `stage.rs` | 813 | one `[stages.<name>]`: model, mode, tools, transitions |
| `regions.rs` | 307 | `[context]` and `[context.regions]` |
| `sections.rs` | 228 | the standalone top-level tables |

Glob re-exports keep the split invisible to every caller and to the test
module, the same way `pipeline/mod.rs` does it. The only edit to any moved
function is `pub(super)` on its signature.

### Verified as a pure move

Not asserted - checked. Every non-comment line of the old file and the new
ones, sorted and compared: 1,159 lines in, 1,159 out, zero difference. The
761 tests are untouched and still pass, which they would not if a parse
order had shifted.
